+++ fakernaija/__init__.py
@@ -1,7 +1,7 @@
 """The fakernaija package provides a suite of tools for generating fake data specific to Nigeria.
 
 This package includes various providers that generate realistic Nigerian data such as names,
-emails, phone numbers, schools, and states.
+emails, phone numbers, schools, states etc.
 
 Modules:
 - providers: Contains the different data providers used to generate fake Nigerian data.